new server class, it could be improved to make it mroe readable.  as well right now vector in sock lsitens is not implemented
