+++ includes/Server.hpp
@@ -15,9 +15,13 @@
 #include <sys/epoll.h>  // epoll
 #include <signal.h>  // signal handling
 #include <stdlib.h>
+#include <stdio.h>
 
 #include "ClientHandler.hpp" 
 
+#define PORT "8694"
+#define BACKLOG 5
+#define BUFFER_SIZE 2048
 /*
 Server Class:
 
@@ -44,25 +48,36 @@ private:
     std::vector<int>            sock_listens;  // to list to multiple ports
     std::map<int, HttpRequest>  ongoingRequests;  // ongoing requests for each client_fd
 
+    static volatile sig_atomic_t	    run;
     // Configuration details can go here.
     // E.g., struct Config or std::map<std::string, std::string> config;
     
-    void    setUpSocket(int port);  // For initializing and binding a socket for a given port
-    int     setUpEpoll(int sock_listen);  // Setting up epoll
-    int     accept_new_client(int sock_listen);  // Handling new client connection
-    void    handle_client_data(int client_fd);  // Handling data from a client
-    
 public:
     Server();  // Constructor might accept configuration file path as an argument
+    ~Server();
+private:
     Server(const Server& other);
     Server& operator=(const Server& other);
-    ~Server();
+    
+public:
+    void    loadConfig(const std::string& configPath);  // Load server configurations from a file
+    void    start();  // Start the server
+    void    stop();  // Stop the server
 
-    void start();  // Start the server
-    void stop();  // Stop the server
+private:
+    // Encapsulate all the helper methods inside private scope
+    int     set_and_bind_sock_listen(int *sock_listen);
+    int     setUpSocket(int* sock_listen);
+    int     setUpEpoll(int sock_listen);
+    int     accept_new_client(int epoll_fd, int sock_listen);
+    void    handle_client_data(int epoll_fd, int client_fd);
+    int     calculate_dynamic_timeout();
+    int     handle_epoll_events(int epoll_fd, int sock_listen);
+    static void	signal_handler(int sig);
 
-    void loadConfig(const std::string& configPath);  // Load server configurations from a file
-
+public:
     friend std::ostream& operator<<(std::ostream& os, const Server & server);
 };
+
+
 #endif
