+++ sources/HttpRequestDELETE.cpp
@@ -4,11 +4,16 @@
 HttpRequestDELETE::HttpRequestDELETE() {
 }
 
-HttpRequestDELETE::HttpRequestDELETE(/*args*/){
+HttpRequestDELETE::HttpRequestDELETE(int tmp_holder) {
+    (void)tmp_holder;
 }
 
-HttpRequestDELETE::HttpRequestDELETE(const HttpRequestDELETE & other){
+HttpRequestDELETE::HttpRequestDELETE(const HttpRequestDELETE & other) : HttpRequestBase(other)
+{
+    (void)other;
 }
 
 HttpRequestDELETE	&HttpRequestDELETE::operator=(const HttpRequestDELETE & other) {
+    (void)other;
+    return *this;
 }
