+++ sources/HttpRequestGET.cpp
@@ -4,11 +4,16 @@
 HttpRequestGET::HttpRequestGET() {
 }
 
-HttpRequestGET::HttpRequestGET(/*args*/){
+HttpRequestGET::HttpRequestGET(int tmp_holder){
+    (void)tmp_holder;
 }
 
-HttpRequestGET::HttpRequestGET(const HttpRequestGET & other){
+HttpRequestGET::HttpRequestGET(const HttpRequestGET & other) : HttpRequestBase(other)
+{
+    (void)other;
 }
 
 HttpRequestGET	&HttpRequestGET::operator=(const HttpRequestGET & other) {
+    (void)other;
+    return *this;
 }
