+++ sources/main.cpp
@@ -12,17 +12,8 @@
 
 #include "Server.hpp"
 
-//Move this to Server class!
-volatile sig_atomic_t	run = 1;
-
-void	signal_handler(int sig)
-{
-	if (sig == SIGINT)
-	run = 0;
-}
 
 int main(void) {
-    signal(SIGINT, signal_handler);  // This handler should set a global flag or a Server member to indicate server should shut down.
 
     Server server;
     server.loadConfig("path_to_config");  // If you have a configuration file.
